atualizando home e login

+++ playOnMatch/back/src/controllers/controllerEncontro.js
@@ -6,20 +6,20 @@ const prisma = new PrismaClient()
 
 const create = async (req, res) => {
     try {
-
-        const encontro = await prisma.encontro.create({
+        //cria novo encontro
+        const newEncontro = await prisma.encontro.create({
             data: req.body
         });
 
+        // Cria um novo registro para o encontro e os usuários relacionados
         const encontroUsuario = await prisma.encontroUsuario.create({
             data: {
-                id_encontro: encontro.id,
+                id_encontro: newEncontro.id,
                 idCriadorPartida: Number(req.params.idCriadorPartida),
                 idParticipantePartida: Number(req.params.idCriadorPartida)
             },
         });
-
-        res.status(200).send({ mensagem: encontroUsuario }).end()
+        res.status(201).send({ mensagem: encontroUsuario }).end()
     } catch (error) {
         console.log(error);
         res.status(500).send({ mensagem: error }).end();
@@ -41,10 +41,14 @@ const novoParticipante = async (req, res) => {
         }
     })
 
+    if (!encontro) {
+        return res.status(404).send({ mensagem: 'Encontro não encontrado' }).end()
+    }
+
     const encontroUsuario = await prisma.encontroUsuario.create({
         data: {
             id_encontro: Number(req.params.idEncontro),
-            idCriadorPartida: encontro.EncontroUsuario[0].idCriadorPartida,
+            idCriadorPartida: encontro?.EncontroUsuario[0]?.idCriadorPartida,
             idParticipantePartida: Number(req.params.idNovoParticipante)
         }
     })
@@ -62,73 +66,104 @@ const deletarParticipante = async (req, res) => {
                 EncontroUsuario: true
             }
         })
-        if (encontro) {
-            encontro.EncontroUsuario.forEach(async (e) => {
-                if (e.idParticipantePartida == req.params.idParticipante) {
-                    await prisma.encontroUsuario.delete({
-                        where: {
-                            id: e.id
-                        }
-                    })
 
-                    res.status(200).send(encontro).end()
-                } else {
-                    res.status(400).send({ mensagem: "vc não pode cancelar a ida de outro participante" }).end()
-                }
-            })
-        } else {
-            res.status(400).send({ mensagem: "encontro não existe" }).end()
+        if (!encontro) {
+            return res.status(400).send({ mensagem: "Encontro não encontrado" }).end()
         }
 
+        encontro.EncontroUsuario.forEach(async (e) => {
+            if (e.idParticipantePartida == req.params.idParticipante) {
+                await prisma.encontroUsuario.delete({
+                    where: {
+                        id: e.id
+                    }
+                })
+
+                res.status(200).send(encontro).end()
+            } else {
+                res.status(400).send({ mensagem: "vc não pode cancelar a ida de outro participante" }).end()
+            }
+        })
 
     } catch (error) {
         console.log(error)
-        res.status(400).send(error).end()
+        res.status(500).send(error).end()
     }
 }
 
 const readAll = async (req, res) => {
-    let encontro = await prisma.encontro.findMany({
-        select: {
-            id: true,
-            descricao: true,
-            data: true,
-            titulo: true,
-            id_local: true,
-            local: {
-                select: {
-                    nome: true,
-                    capacidade: true,
-                    endereco: true
-                }
-            },
-            EncontroUsuario: {
-                select: {
-                    idCriadorPartida: true,
-                    idParticipantePartida: true
+
+    try {
+        let encontro = await prisma.encontro.findMany({
+            select: {
+                id: true,
+                descricao: true,
+                data: true,
+                titulo: true,
+                id_local: true,
+                local: {
+                    select: {
+                        nome: true,
+                        capacidade: true,
+                        endereco: true
+                    }
+                },
+                EncontroUsuario: {
+                    select: {
+                        idParticipante: {
+                            select: {
+                                nome: true,
+                                id: true
+                            }
+                        },
+                        idCriador: {
+                            select: {
+                                id: true,
+                                nome: true,
+                                criadorPartida:true
+                            }
+                        }
+                    }
                 }
             }
-        }
-    })
-    res.status(200).json(encontro).end()
+        })
+        res.status(200).json(encontro).end()
+    } catch (error) {
+        console.log(error)
+        res.status(500).send({ mensagem: "Erro ao buscar encontros" }).end()
+    }
+
 }
 
 const readOne = async (req, res) => {
     let encontro = await prisma.encontro.findUnique({
         where: {
             id: Number(req.params.id)
+        },
+        select:{
+            EncontroUsuario:{
+                select:{
+                    idParticipante:{
+                        select:{
+                            id:true,
+                            nome:true
+                        }
+                    },
+                    idCriador:{
+                        select:{
+                            id:true,
+                            nome:true
+                        }
+                    }
+                }
+            }
         }
     })
-    if (encontro != null) {
-        res.status(200).json(encontro).end()
-    } else {
-        res.status(200).send({
-            mensagem: "erro"
-        }).end()
-    }
+    encontro ? res.status(200).json(encontro).end() : res.status(404).send({ mensagem: "Encontro não encontrado" }).end()
 }
 
 const update = async (req, res) => {
+
     try {
         let encontro = await prisma.encontro.update({
             where: {
@@ -141,7 +176,7 @@ const update = async (req, res) => {
         }).end()
     } catch (error) {
         console.log(error)
-        res.status(200).send({
+        res.status(404).send({
             mensagem: `Erro ${error.code}, id do encontro não foi achado`
         }).end()
     }
@@ -149,25 +184,35 @@ const update = async (req, res) => {
 
 const del = async (req, res) => {
     try {
-        let encontro = await prisma.encontro.delete({
+        const encontro = await prisma.encontro.findUnique({
             where: {
                 id: Number(req.params.id)
             }
         })
+
+        if (!encontro) {
+            return res.status(404).send({
+                mensagem: "Encontro não encontrado"
+            }).end()
+        }
+
+        await prisma.encontro.delete({
+            where: {
+                id: Number(req.params.id)
+            }
+        })
+
         res.status(200).send({
-            mensagem: `Encontro de id ${encontro.id} foi excluido com sucesso`
+            mensagem: `Encontro de id ${encontro.id} foi excluído com sucesso`
         }).end()
     } catch (err) {
-        res.status(200).send({
-            mensagem: "erro"
+        console.log(err)
+        res.status(500).send({
+            mensagem: "Ocorreu um erro ao excluir o encontro"
         }).end()
     }
 }
 
-
-// const postEncontroUsuario = async (req, res) => {
-
-// }
 module.exports = {
     create,
     novoParticipante,
